django-ninja-example/pyproject.toml
[project]
name = "django-ninja-example"
version = "0.1.0"
requires-python = "==3.14.*"
dependencies = [
    "faker==39.0.0",
    "uvicorn[standard]",
    "Django==6.0",
    "orjson==3.11.5",
    "psycopg[binary]==3.3.2",
    # library under test
    "django-ninja>=1.5.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"

[tool.uv]
package = false
